CSV to XML script.
Initial version of the csv to xml script for conversion of GRIB/BUFR
tables. Script reads all csv files in the directory (not in subdirectories), converts all csv to xml, keeping the file name. The root element is set to “dataroot”, element of each record is the file name, e.g. “BUFRCREX_33_0_0_CodeFlag_en”, as it was in previous xml exports. These are simply configurable.